tests/submission/W3C/domhtml/HTMLDocument17.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/dynamic-markup-insertion.html#dom-document-open" rel="help">
<link rel="author" title="Curt Arnold">
</head>
<body id="TEST-BODY"> 
<p>
Clears the current document using HTMLDocument.open immediately followed by close.
</p>
<iframe id='iframe' src='/support/html/sample.html'>
</iframe>

<script>
var domtest = async_test("Clears the current document using HTMLDocument.open immediately followed by close.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var doc;
      var bodyElem;
      var bodyChild;
      doc = window.document.getElementById('iframe').contentWindow.document;
      doc.open();
      doc.close();
      bodyElem = doc.body;

      
	if(
	
	(bodyElem != null)

	) {
	bodyChild = bodyElem.firstChild;

      assert_equals(bodyChild, null);
    
	}
	
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
